Add a separate method for writing output files in addition to printStatistics.
Separating statistics from output files will allow us for example
to write output files in parallel (not yet implemented).
Furthermore, the RestartAlgorithm will not call the method for writing
output files on all intermediate analyses
(i.e., those that are not the last analysis of a CPAchecker run).
This will improve performance and create less useless output files.

This commit does not actually change any statistics implementation
to use the new method, so for now nothing changes.
In the future, we can start migrating statistics of components
to actually separate statistics and output files.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26919 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c