removing ARGReplayAlgorithm.
The algorithm was used years ago to develop a combination of predicate analyses,
e.g., first linear, then bitprecise with reusing formulas from the first analysis.
The overall outcome was that bitprecise analysis is already fast enough and
the combined analysis is not needed.

As there is a possibility to reuse the ARG in the default RestartAlgorithm,
it maybe should better be integrated there, if needed.

We keep the ARGReplayCPA, as it is independent of other components and might
be useful in some cases. However, for correct usage, we need to find a way
to not close the first CPAs, including the SMT solver, to be able to access
all information from the ARG. This is low priority for now, thus ignoring it.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26933 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c