Cleanup test sets for memory-safety
It is important not to add arbitrary options to test sets.
If the two options should have been used in SV-COMP,
they would need to have been added to the SV-COMP config.

Similarly for the integration test. If these two options are useful in
general, they should be added to the proper config file such that
everyone can benefit from them, not only the BuildBot.
Furthermore, it does not make sense to test an old config of SV-COMP'16
in the BuildBot.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26963 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c